refactor(radarr_compat): split 834-line file into 5-module subdir — Phase 4 complete
  radarr_compat/
    mod.rs      37   router glue + auth middleware + pub re-exports
    types.rs   127   RadarrMovie + 4 nested types (Image, Ratings,
                      RatingValue, MovieFile) + RadarrRootFolder +
                      4 request DTOs (AddMovieBody, AddMovieOptions,
                      UpdateMovieBody, RadarrCommandBody)
    system.rs   92   system_status, quality_profiles, root_folders,
                      list_tags, list_download_clients, create_tag
    movie.rs   322   movie_lookup, list_movies, get_movie, add_movie,
                      update_movie, execute_command
    helpers.rs 295   lookup_by_tmdb_id, build_radarr_movie_*,
                      map_status, default_ratings, build_stub_movie

Same pattern as sonarr_compat split (commit 5a0440f):
  - 6 helpers bumped to pub(super) for sibling access
  - RadarrRootFolder fields made pub for cross-module construction
  - super::library → super::super::library for the one auto-search
    call-site inside add_movie

Phase 4 complete. Summary:
  Before:  sonarr_compat.rs (1381 lines) + radarr_compat.rs (921 lines)
  After:   handlers/arr_shared.rs (123 lines — 6 shared DTOs)
           handlers/sonarr_compat/ (1336 lines across 5 files)
           handlers/radarr_compat/ (873 lines across 5 files)
  Net: -10 lines with clearer structure + single source of truth for
       SystemStatus / QualityProfile / Tag / DownloadClientEntry /
       TagBody / LookupQuery (was duplicated in both shims).

External handler paths unchanged. 601/601 tests pass.
